Project 4 clean up

+++ project_04/map.h
@@ -1,14 +1,14 @@
 #ifndef UNO_PROJECT04_MAP_H_
 #define UNO_PROJECT04_MAP_H_
 
-#include "array.h"
+#include <array.h>
 
 namespace uno {
 
 constexpr array<int, 3> map(const array<int, 3>& sensorValues) noexcept {
-  return {static_cast<int>(sensorValues[0]/4),
-          static_cast<int>(sensorValues[1]/4),
-          static_cast<int>(sensorValues[2]/4)};
+  return {sensorValues[0]/4,
+          sensorValues[1]/4,
+          sensorValues[2]/4};
 }
 
 }  // namespace uno
